examples/inference/stable-diffusion-python-onnx/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

transformers==4.40.1
diffusers==0.27.2
ftfy==6.2.0
optimum[onnxruntime]==1.19.1
pillow==10.3.0
